refactor(clinics): move clinic equipment methods inside clinic service
