+++ src/main/java/com/medspace/application/usecase/clinic/GetAllClinicsUseCase.java
@@ -1,7 +1,6 @@
 package com.medspace.application.usecase.clinic;
 
 import com.medspace.application.service.ClinicAvailabilityService;
-import com.medspace.application.service.ClinicEquipmentService;
 import com.medspace.application.service.ClinicService;
 import com.medspace.domain.model.Clinic;
 import com.medspace.infrastructure.dto.clinic.ClinicQueryDTO;
@@ -19,8 +18,6 @@ public class GetAllClinicsUseCase {
     @Inject
     ClinicService clinicService;
     @Inject
-    ClinicEquipmentService clinicEquipmentService;
-    @Inject
     ClinicAvailabilityService clinicAvailabilityService;
 
     public List<GetClinicDTO> execute(ClinicQueryDTO queryFilterDTO) {
@@ -36,8 +33,8 @@ public class GetAllClinicsUseCase {
 
             List<GetClinicEquipmentDTO> equipmentDTOs =
                     queryFilterDTO.getIncludeEquipments()
-                            ? clinicEquipmentService.getEquipmentsByClinicId(clinic.getId())
-                                    .stream().map(GetClinicEquipmentDTO::new).toList()
+                            ? clinicService.getEquipmentsByClinicId(clinic.getId()).stream()
+                                    .map(GetClinicEquipmentDTO::new).toList()
                             : null;
 
             List<GetClinicAvailabilityDTO> availabilityDTOs =
