+++ src/main/java/com/medspace/application/usecase/clinic/GetClinicByIdUseCase.java
@@ -2,7 +2,6 @@ package com.medspace.application.usecase.clinic;
 
 import java.util.List;
 import com.medspace.application.service.ClinicAvailabilityService;
-import com.medspace.application.service.ClinicEquipmentService;
 import com.medspace.application.service.ClinicService;
 import com.medspace.domain.model.Clinic;
 import com.medspace.infrastructure.dto.clinic.ClinicQueryDTO;
@@ -19,8 +18,6 @@ public class GetClinicByIdUseCase {
     @Inject
     ClinicService clinicService;
     @Inject
-    ClinicEquipmentService clinicEquipmentService;
-    @Inject
     ClinicAvailabilityService clinicAvailabilityService;
 
     public GetClinicDTO execute(Long id, ClinicQueryDTO queryFilterDTO) {
@@ -36,10 +33,11 @@ public class GetClinicByIdUseCase {
                                 .map(GetClinicPhotoDTO::new).toList()
                         : null;
 
-        List<GetClinicEquipmentDTO> equipmentDTOs = queryFilterDTO.getIncludeEquipments()
-                ? clinicEquipmentService.getEquipmentsByClinicId(id).stream()
-                        .map(GetClinicEquipmentDTO::new).toList()
-                : null;
+        List<GetClinicEquipmentDTO> equipmentDTOs =
+                queryFilterDTO.getIncludeEquipments()
+                        ? clinicService.getEquipmentsByClinicId(id).stream()
+                                .map(GetClinicEquipmentDTO::new).toList()
+                        : null;
 
         List<GetClinicAvailabilityDTO> availabilityDTOs = queryFilterDTO.getIncludeAvailabilities()
                 ? clinicAvailabilityService.getAvailabilitiesByClinicId(id).stream()
