+++ src/main/java/com/medspace/application/usecase/clinicEquipment/CreateClinicEquipmentUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.clinicEquipment;
 
-import com.medspace.application.service.ClinicEquipmentService;
+import com.medspace.application.service.ClinicService;
 import com.medspace.domain.model.ClinicEquipment;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
@@ -8,9 +8,9 @@ import jakarta.inject.Inject;
 @ApplicationScoped
 public class CreateClinicEquipmentUseCase {
     @Inject
-    ClinicEquipmentService clinicEquipmentService;
+    ClinicService clinicService;
 
     public ClinicEquipment execute(ClinicEquipment clinicEquipment) {
-        return clinicEquipmentService.createEquipment(clinicEquipment);
+        return clinicService.createClinicEquipment(clinicEquipment);
     }
 }
