+++ src/main/java/com/medspace/application/usecase/clinicEquipment/DeleteClinicEquipmentByIdUseCase.java
@@ -1,6 +1,5 @@
 package com.medspace.application.usecase.clinicEquipment;
 
-import com.medspace.application.service.ClinicEquipmentService;
 import com.medspace.application.service.ClinicService;
 import com.medspace.domain.model.ClinicEquipment;
 import io.quarkus.security.ForbiddenException;
@@ -11,19 +10,17 @@ import jakarta.transaction.Transactional;
 @ApplicationScoped
 public class DeleteClinicEquipmentByIdUseCase {
     @Inject
-    ClinicEquipmentService clinicEquipmentService;
-    @Inject
     ClinicService clinicService;
 
     @Transactional
     public void execute(Long id, Long userId) {
-        ClinicEquipment clinicEquipment = clinicEquipmentService.getEquipmentById(id);
+        ClinicEquipment clinicEquipment = clinicService.getClinicEquipmentById(id);
         Long clinicId = clinicEquipment != null ? clinicEquipment.getClinic().getId() : null;
         Boolean isOwner = clinicService.validateClinicOwnership(clinicId, userId);
         if (!isOwner) {
             throw new ForbiddenException("Delete unauthorized");
         }
 
-        clinicEquipmentService.deleteEquipmentById(id);
+        clinicService.deleteClinicEquipmentById(id);
     }
 }
