+++ src/main/java/com/medspace/application/usecase/clinicEquipment/GetEquipmentsByClinicIdUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.clinicEquipment;
 
-import com.medspace.application.service.ClinicEquipmentService;
+import com.medspace.application.service.ClinicService;
 import com.medspace.domain.model.ClinicEquipment;
 import com.medspace.infrastructure.dto.clinic.GetClinicEquipmentDTO;
 import jakarta.enterprise.context.ApplicationScoped;
@@ -12,10 +12,10 @@ import java.util.List;
 @ApplicationScoped
 public class GetEquipmentsByClinicIdUseCase {
     @Inject
-    ClinicEquipmentService clinicEquipmentService;
+    ClinicService clinicService;
 
     public List<GetClinicEquipmentDTO> execute(Long id) {
-        List<ClinicEquipment> clinicEquipments = clinicEquipmentService.getEquipmentsByClinicId(id);
+        List<ClinicEquipment> clinicEquipments = clinicService.getEquipmentsByClinicId(id);
         List<GetClinicEquipmentDTO> clinicEquipmentDTOs = new ArrayList<>();
 
         for (ClinicEquipment clinicEquipment : clinicEquipments) {
